Fix mac build, at end of render, refresh the grid

+++ xLights/RenderFan.cpp
@@ -97,7 +97,7 @@ void RgbEffects::RenderFan(int center_x, int center_y, int start_radius, int end
         std::swap(radius1, radius2);
     }
 
-    double a_const = radius2 / ToRadians(std::abs(blade_angle));
+    double a_const = radius2 / ToRadians(std::abs((double)blade_angle));
 
     for( int blade = 0; blade < num_blades; blade++ )
     {
